requirements.txt
mesa
networkx
numpy
pandas
scipy
matplotlib
jmetalpy